Definicion de funcion middleware para obtener el usuario que hace cada peticion en las rutas que inician con /api/blogs

+++ controllers/blogs.js
@@ -3,8 +3,6 @@ const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const jwt = require('jsonwebtoken')
-
 blogRouter.get('/', async (req, res) => {
     const blogs = await Blog.find({}).populate('user', { blogs: 0 })
 
@@ -13,20 +11,19 @@ blogRouter.get('/', async (req, res) => {
 
 blogRouter.post('/', async (req, res) => {
     const body = req.body
+    const userAuth = req.user
 
-    const userToken = jwt.verify(req.token, process.env.SECRET)
-
-    if (userToken.userid) {
+    if (userAuth.userid) {
         const newBlog = new Blog({
             author: body.author,
             likes: body.likes,
             title: body.title,
             url: body.url,
-            user: userToken.userid // id del usuario creador del blog
+            user: userAuth.userid // id del usuario creador del blog
         })
         const blog = await newBlog.save()
 
-        const user = await User.findById(userToken.userid)
+        const user = await User.findById(userAuth.userid)
         user.blogs = user.blogs.concat(blog.id) // concatenacion de los blogs asociados a un usuario
         await user.save()
 
@@ -37,16 +34,17 @@ blogRouter.post('/', async (req, res) => {
 })
 
 blogRouter.delete('/:id', async (req, res) => {
-    const deletedBlog = await Blog.findByIdAndDelete(req.params.id)
+    const userAuth = req.user
 
-    const userToken = jwt.verify(req.token, process.env.SECRET)
+    const blogToDelete = await Blog.findById(req.params.id)
 
-    if (userToken.userid) {
-        if (deletedBlog) {
-            if (userToken.userid === deletedBlog.user.toString()) {
+    if (userAuth.userid) {
+        if (blogToDelete) {
+            if (userAuth.userid === blogToDelete.user.toString()) {
+                await Blog.findByIdAndDelete(req.params.id)
                 return res.status(204).end()
             }
-            return res.status(401).json({ error: "the specified user cannot delete the blog" })
+            return res.status(401).json({ error: "the user is not authorized to delete the blog" })
         }
 
         return res.status(404).json({ error: 'the specified blog doesn\'t exist' })
